@@ -1,33 +1,35 @@
 {
-  "name": "helloworld-react",
+  "name": "sample",
   "version": "1.0.0",
-  "description": "In this sample we are going to setup a basic hello world example with React.",
+  "description": "In this sample we are going to setup the basic plumbing to \"build\" our project and launch it in a dev server.",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server --inline --hot --open",
-    "build": "webpack"
+    "start": "webpack-dev-server  --mode development --inline --hot --open",
+    "build": "webpack  --mode development",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "Lemoncode and Front End Master Students",
-  "license": "MIT",
-  "dependencies": {
-    "bootstrap": "~4.0.0",
-    "react": "~16.2.0",
-    "react-dom": "~16.2.0"
-  },
+  "author": "",
+  "license": "ISC",
   "devDependencies": {
-    "@types/react": "~16.0.36",
-    "@types/react-dom": "~16.0.3",
-    "awesome-typescript-loader": "^3.4.1",
+    "@types/react": "^16.3.8",
+    "@types/react-dom": "^16.0.5",
+    "awesome-typescript-loader": "^5.0.0",
     "babel-core": "^6.26.0",
     "babel-preset-env": "^1.6.1",
-    "css-loader": "^0.28.9",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^1.1.6",
-    "html-webpack-plugin": "^2.30.1",
-    "style-loader": "^0.20.1",
-    "typescript": "^2.7.1",
-    "url-loader": "^0.6.2",
-    "webpack": "^3.10.0",
-    "webpack-dev-server": "^2.11.1"
+    "css-loader": "^0.28.11",
+    "file-loader": "^1.1.11",
+    "html-webpack-plugin": "^3.2.0",
+    "mini-css-extract-plugin": "^0.4.0",
+    "style-loader": "^0.20.3",
+    "typescript": "^2.8.1",
+    "url-loader": "^1.0.1",
+    "webpack": "^4.5.0",
+    "webpack-cli": "^2.0.14",
+    "webpack-dev-server": "3.1.0"
+  },
+  "dependencies": {
+    "bootstrap": "^4.1.0",
+    "react": "^16.3.1",
+    "react-dom": "^16.3.1"
   }
 }
