@@ -1,7 +1,7 @@
 let path = require('path');
-let webpack = require('webpack');
 let HtmlWebpackPlugin = require('html-webpack-plugin');
-let ExtractTextPlugin = require('extract-text-webpack-plugin');
+let MiniCssExtractPlugin = require('mini-css-extract-plugin');
+let webpack = require('webpack');
 
 let basePath = __dirname;
 
@@ -10,7 +10,6 @@ module.exports = {
   resolve: {
     extensions: ['.js', '.ts', '.tsx']
   },
-
   entry: [
     './main.ts',
     '../node_modules/bootstrap/dist/css/bootstrap.css'
@@ -19,9 +18,7 @@ module.exports = {
     path: path.join(basePath, 'dist'),
     filename: 'bundle.js'
   },
-
   devtool: 'source-map',
-
   devServer: {
     contentBase: './dist', // Content base
     inline: true, // Enable watch and live reload
@@ -29,9 +26,9 @@ module.exports = {
     port: 8080,
     stats: 'errors-only'
   },
-
   module: {
-    rules: [{
+    rules: [
+      {
         test: /\.(ts|tsx)$/,
         exclude: /node_modules/,
         loader: 'awesome-typescript-loader',
@@ -40,32 +37,8 @@ module.exports = {
         },
       },
       {
-        test: /\.css$/,
-        include: /node_modules/,
-        loader: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: {
-            loader: 'css-loader',
-          },
-        }),
-      },
-      // Loading glyphicons => https://github.com/gowravshekar/bootstrap-webpack
-      // Using here url-loader and file-loader
-      {
-        test: /\.(woff|woff2)(\?v=\d+\.\d+\.\d+)?$/,
-        loader: 'url-loader?limit=10000&mimetype=application/font-woff'
-      },
-      {
-        test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/,
-        loader: 'url-loader?limit=10000&mimetype=application/octet-stream'
-      },
-      {
-        test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
-        loader: 'url-loader?limit=10000&mimetype=image/svg+xml'
-      },
-      {
-        test: /\.eot(\?v=\d+\.\d+\.\d+)?$/,
-        loader: 'file-loader'
+        test: /\.css$/,        
+        use: [MiniCssExtractPlugin.loader, "css-loader"]
       },
       {
         test: /\.(png|jpg|gif|svg)$/,
@@ -74,19 +47,18 @@ module.exports = {
           name: 'assets/img/[name].[ext]?[hash]'
         }
       },
-    ]
+    ],
   },
   plugins: [
-    // Generate index.html in /dist => https://github.com/ampedandwired/html-webpack-plugin
+    //Generate index.html in /dist => https://github.com/ampedandwired/html-webpack-plugin
     new HtmlWebpackPlugin({
-      filename: 'index.html', // Name of file in ./dist/
-      template: 'index.html', // Name of template in ./src
-      hash: true
+      filename: 'index.html', //Name of file in ./dist/
+      template: 'index.html', //Name of template in ./src
+      hash: true,
     }),
-    new ExtractTextPlugin({
-      filename: '[chunkhash].[name].css',
-      disable: false,
-      allChunks: true,
+    new MiniCssExtractPlugin({
+      filename: "[name].css",
+      chunkFilename: "[id].css"
     }),
-  ]
-}
+  ],
+};
