@@ -1,6 +1,7 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import {HelloComponent} from './hello';
+
+import { HelloComponent } from './hello';
 
 ReactDOM.render(
   <HelloComponent userName="John" />,
