@@ -1,8 +1,8 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import {App} from './app';
+import { App } from './app';
 
 ReactDOM.render(
-  <App />,
+  <App/>,
   document.getElementById('root')
 );
