@@ -1,6 +1,4 @@
 import * as React from 'react';
-import {Fragment} from 'react';
-
 
 interface Props {
   userName : string;
@@ -9,9 +7,9 @@ interface Props {
 
 export const NameEditComponent = (props : Props) => {
   return (
-    <Fragment>
+    <>
       <label>Update name:</label>
       <input value={props.userName} onChange={props.onChange}/>
-    </Fragment>
+    </>
   );
 }
