@@ -1,10 +1,6 @@
 import * as React from 'react';
 
-interface Props {
-  userName : string;
-}
-
-export const HelloComponent = (props : Props) => {
+export const HelloComponent = (props: {userName : string}) => {
   return (
     <h2>Hello user: {props.userName} !</h2>
   );
