@@ -1,7 +1,8 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import {App} from './app';
+import { App } from './app';
 
 ReactDOM.render(
-  <App/>
-  , document.getElementById('root'));
+  <App/>,
+  document.getElementById('root')
+);
