@@ -1,17 +1,19 @@
 import * as React from 'react';
+import {Fragment} from 'react';
+
 
 interface Props {
-  editingUserName : string;
-  onEditingNameUpdated : (newEditingName : string) => any;
-  onNameUpdateRequest : () => void;
+    editingUserName : string;
+    onEditingNameUpdated : (newEditingName : string) => void;
+    onNameUpdateRequest : () => void;  
 }
 
-export const NameEditComponent = (props: Props) => {
-    return (
-      <div>
-        <label>Update Name:</label>
-        <input value={props.editingUserName} onChange={(event : any) => props.onEditingNameUpdated(event.target.value)}/>
-        <input type="submit" value="Change" className="btn btn-default" onClick={props.onNameUpdateRequest} />
-      </div>
-    );
-}
+  
+export const NameEditComponent = (props : Props) =>
+  <div>
+      <label>Update Name:</label>
+      <input value={props.editingUserName}
+        onChange={(e) : void => props.onEditingNameUpdated((e.target as HTMLInputElement).value)} />
+
+      <button className="btn btn-default" onClick={props.onNameUpdateRequest}>Change</button>
+  </div>
