@@ -6,17 +6,14 @@ interface Props {
 }
 
 export const ColorDisplayer = (props : Props) => {
-  // `rgb(${props.color.red},${props.color.green}, ${props.color.blue}) })`
-  // 'rgb(' + props.color.red + ', 40, 80)'
-  var divStyle = {
-    width: '120px',
-    height: '80px',
+  const divStyle = {
+    width: '11rem',
+    height: '7rem',
     backgroundColor: `rgb(${props.color.red},${props.color.green}, ${props.color.blue})`
   };
 
-
   return (
-    <div style={divStyle}>    
+    <div style={divStyle}>
     </div>
   );
 }
