@@ -1,18 +1,21 @@
 import * as React from 'react';
-import { Color } from './color';
 import { ColorSliderComponent } from './colorslider';
+
+import { Color } from './color'
 
 interface Props {
   color: Color;
   onColorUpdated: (color: Color) => void;
 }
 
-const updateColor = (props: Props, colorId : keyof Color) => (value) => {
+
+const updateColor = (props: Props, colorId: keyof Color) => (value) => {
   props.onColorUpdated({
     ...props.color,
     [colorId]: value
   });
 };
+
 
 
 export const ColorPicker = (props: Props) => {
@@ -22,7 +25,6 @@ export const ColorPicker = (props: Props) => {
         value={props.color.red}
         onValueUpdated={updateColor(props, 'red')}
       />
-      {props.color.red}
       <br />
       <ColorSliderComponent
         value={props.color.green}
@@ -31,7 +33,7 @@ export const ColorPicker = (props: Props) => {
       <br />
       <ColorSliderComponent
         value={props.color.blue}
-        onValueUpdated={updateColor(props, 'blue')}
+        onValueUpdated={updateColor(props, 'blue')}        
       />
       <br />
     </div>
