@@ -3,5 +3,6 @@ import * as ReactDOM from 'react-dom';
 import {App} from './app';
 
 ReactDOM.render(
-  <App/>
-  , document.getElementById('root'));
+  <App/>,
+  document.getElementById('root')
+);
