@@ -1,17 +1,17 @@
-import {MemberEntity} from '../model/member';
+import { MemberEntity } from '../model/member';
 
-var	MembersMockData : MemberEntity[] =
-	[
-		{
+var MembersMockData: MemberEntity[] =
+  [
+    {
       id: 1457912,
       login: "brauliodiez",
       avatar_url: "https://avatars.githubusercontent.com/u/1457912?v=3"
-		},
+    },
     {
       id: 4374977,
       login: "Nasdan",
       avatar_url: "https://avatars.githubusercontent.com/u/4374977?v=3"
     }
-	];
+  ];
 
-  export default MembersMockData;
+export default MembersMockData;
