@@ -1,24 +1,15 @@
 import * as React from 'react';
-import {MemberEntity} from './model/member';
+import { MemberEntity } from './model/member';
 
-
-interface Props  {
-  member : MemberEntity;
-}
-
-export const MemberRow = (props: Props) => {
-
-     return (
-       <tr>
-         <td>
-           <img src={props.member.avatar_url} className="avatar"/>
-         </td>
-         <td>
-           <span>{props.member.id}</span>
-         </td>
-         <td>
-           <span>{props.member.login}</span>
-         </td>
-       </tr>
-     );
-}
+export const MemberRow = (props: { member: MemberEntity }) =>
+  <tr>
+    <td>
+      <img src={props.member.avatar_url} style={{ maxWidth: '10rem' }} />
+    </td>
+    <td>
+      <span>{props.member.id}</span>
+    </td>
+    <td>
+      <span>{props.member.login}</span>
+    </td>
+  </tr>
